Move distributions into their own submodules
A new class DistributionInstaller bundles everything needed to install a certain
distribution. Every supported distribution inherits from this to specify how it
is to be installed.

To break circular imports the most important types CommandLineArguments and
DistributionInstaller, the latter being added in this commit, are moved to a
module backend.py. All the functions and classes they depend on are moved as
well.

To forestall further circular imports the main mkosi module (mkosi.__init__)
accesses the distribution installers via a lazy
object (LazySupportedDistrosDict)
